app_line/requirements.txt
Flask==1.1.2
line-bot-sdk==2.0.1
google-api-python-client==2.26.1